[Refactor] UseInfiniteScroll 수정 중 - useEffect 내부 observe & unobserve 로직 완성
